+++ index.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="./js/main.jsx"></script>
+    <script type="module" src="/src/main.jsx"></script>
   </body>
 </html>
