+++ src/main.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App";
+import App from "./components/App";
 // import "./index.css";
-import "../scss/styles.scss";
-import * as bootstrap from "bootstrap";
+import "./scss/styles.scss";
+// import * as bootstrap from "bootstrap";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
