+++ vite.config.js
@@ -5,7 +5,6 @@ import path from "path";
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
-  root: path.resolve((__dirname, "src")),
   resolve: {
     alias: {
       "~bootstrap": path.resolve((__dirname, "node_modules/bootstrap")),
